# api/openapi/user.openapi.yaml
openapi: 3.0.1
info:
  title: MyCloudDrive User API
  description: 用户域接口契约（Go 风格命名）。
  version: 1.0.0
tags:
  - name: 用户管理
  - name: 用户管理/认证管理
  - name: 用户管理/用户工作空间
  - name: 用户管理/用户数据
  - name: 认证管理
  - name: 用户工作空间
paths:
  /apis/auth/logout:
    post:
      summary: 登出
      deprecated: false
      description: ''
      operationId: logout
      tags:
        - 用户管理/认证管理
        - 认证管理
      parameters: []
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
  /apis/auth/login:
    post:
      summary: 登录
      deprecated: false
      description: ''
      operationId: doLogin
      tags:
        - 用户管理/认证管理
        - 认证管理
      parameters:
        - name: Authorization
          in: header
          description: ''
          required: false
          example: >-
            Bearer
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIwMWpydmdzOTQzcTBmNDNoMGFhNW1qZGUweSIsIndzIjoid3NfMDFqcnZnczk0M3EwZjQzaDBhYTVtamRlMHlfcGVyc29uYWwiLCJ1biI6ImFkbWluIiwiaWF0IjoxNzc3MDI2OTY0LCJleHAiOjE3Nzc2MzE3NjR9.Go6A0dP_wASdk1EWCBSzB86VyNy87ysQfrjriFEgnKg
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCmd'
            example:
              username: admin
              password: admin
              isRemember: true
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
  /apis/user/workspaces:
    get:
      summary: 获取当前用户可访问工作空间列表
      deprecated: false
      description: ''
      operationId: listUserWorkspaces
      tags:
        - 用户管理/用户工作空间
        - 用户工作空间
      parameters:
        - name: Authorization
          in: header
          description: ''
          required: false
          example: >-
            Bearer
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIwMWpydmdzOTQzcTBmNDNoMGFhNW1qZGUweSIsIndzIjoid3NfMDFqcnZnczk0M3EwZjQzaDBhYTVtamRlMHlfcGVyc29uYWwiLCJ1biI6ImFkbWluIiwiaWF0IjoxNzc3MDI2OTY0LCJleHAiOjE3Nzc2MzE3NjR9.Go6A0dP_wASdk1EWCBSzB86VyNy87ysQfrjriFEgnKg
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            example: {}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        workspaceType:
                          type: string
                          enum:
                            - personal
                            - org
                        role:
                          type: string
                          enum:
                            - owner
                            - admin
                            - member
                        isDefault:
                          type: boolean
          headers: {}
      security: []
    post:
      summary: 创建组织工作空间
      deprecated: false
      description: ''
      operationId: createOrgWorkspace
      tags:
        - 用户管理/用户工作空间
        - 用户工作空间
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: 工作空间名称
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      workspaceType:
                        type: string
                        enum:
                          - org
          headers: {}
      security: []
  /apis/user/default-workspace/{workspaceId}:
    put:
      summary: 设置默认工作空间
      deprecated: false
      description: ''
      operationId: setDefaultWorkspace
      tags:
        - 用户管理/用户工作空间
        - 用户工作空间
      parameters:
        - name: workspaceId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      defaultWorkspaceId:
                        type: string
          headers: {}
      security: []
  /apis/user/workspaces/{workspaceId}/members:
    get:
      summary: 获取工作空间成员
      deprecated: false
      description: ''
      operationId: listWorkspaceMembers
      tags:
        - 用户管理/用户工作空间
        - 用户工作空间
      parameters:
        - name: workspaceId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                        role:
                          type: string
                          enum:
                            - owner
                            - admin
                            - member
                        joinedAt:
                          type: string
                          format: date-time
          headers: {}
      security: []
    post:
      summary: 添加工作空间成员
      deprecated: false
      description: ''
      operationId: addWorkspaceMember
      tags:
        - 用户管理/用户工作空间
        - 用户工作空间
      parameters:
        - name: workspaceId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - role
              properties:
                userId:
                  type: string
                role:
                  type: string
                  enum:
                    - admin
                    - member
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      workspaceId:
                        type: string
                      userId:
                        type: string
                      role:
                        type: string
          headers: {}
      security: []
  /apis/user/workspaces/{workspaceId}/members/{userId}:
    delete:
      summary: 移除工作空间成员
      deprecated: false
      description: ''
      operationId: removeWorkspaceMember
      tags:
        - 用户管理/用户工作空间
        - 用户工作空间
      parameters:
        - name: workspaceId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
        - name: userId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      workspaceId:
                        type: string
                      userId:
                        type: string
          headers: {}
      security: []
  /apis/user/transfer/setting:
    get:
      summary: 获取用户传输配置
      deprecated: false
      description: ''
      operationId: getUserTransferSetting
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysUserTransferSetting'
          headers: {}
      security: []
    put:
      summary: 修改用户传输配置
      deprecated: false
      description: ''
      operationId: updateUserTransferSetting
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserTransferSettingEditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysUserTransferSetting'
          headers: {}
      security: []
  /apis/user/password:
    put:
      summary: 修改密码
      deprecated: false
      description: ''
      operationId: resetPassword
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordEditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
  /apis/user/info:
    get:
      summary: 获取用户详细信息
      deprecated: false
      description: ''
      operationId: getDetail
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysUserResponse'
          headers: {}
      security: []
    put:
      summary: 编辑用户
      deprecated: false
      description: ''
      operationId: editUserInfo
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEditInfoRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
  /apis/user/forget-password:
    put:
      summary: 忘记密码-修改密码
      deprecated: false
      description: ''
      operationId: checkForgetPasswordCode
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordForgetEditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
  /apis/user/register:
    post:
      summary: 注册用户
      deprecated: false
      description: ''
      operationId: register
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
            example:
              username: myCloudDriver
              password: myCloudDriver
              confirmPassword: myCloudDriver
              email: is8k1z.rvl54@sina.com
              nickname: 出雨涵
              avatar: https://avatars.githubusercontent.com/u/21450521
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
  /apis/user/forget-password/code/{mail}:
    get:
      summary: 忘记密码-发送验证码
      deprecated: false
      description: ''
      operationId: sendForgetPasswordCodeByMail
      tags:
        - 用户管理/用户数据
        - 用户管理
      parameters:
        - name: mail
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
          headers: {}
      security: []
components:
  schemas:
    ResultObject:
      type: object
      properties:
        data: {}
        code:
          type: integer
          format: int32
        msg:
          type: string
    UserTransferSettingEditRequest:
      type: object
      properties:
        downloadLocation:
          type: string
          minLength: 1
        isDefaultDownloadLocation:
          type: integer
          format: int32
          maximum: 1
          minimum: 0
        downloadSpeedLimit:
          type: integer
          format: int32
          maximum: 200
        concurrentUploadQuantity:
          type: integer
          format: int32
          maximum: 3
        concurrentDownloadQuantity:
          type: integer
          format: int32
          maximum: 3
        chunkSize:
          type: integer
          format: int64
      required:
        - chunkSize
        - concurrentDownloadQuantity
        - concurrentUploadQuantity
        - downloadLocation
        - downloadSpeedLimit
        - isDefaultDownloadLocation
    ResultSysUserTransferSetting:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SysUserTransferSetting'
        code:
          type: integer
          format: int32
        msg:
          type: string
    SysUserTransferSetting:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        userId:
          type: string
        downloadLocation:
          type: string
        isDefaultDownloadLocation:
          type: integer
          format: int32
        downloadSpeedLimit:
          type: integer
          format: int32
        concurrentUploadQuantity:
          type: integer
          format: int32
        concurrentDownloadQuantity:
          type: integer
          format: int32
        chunkSize:
          type: integer
          format: int64
    PasswordEditRequest:
      type: object
      properties:
        oldPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 1
        confirmPassword:
          type: string
          minLength: 1
      required:
        - confirmPassword
        - newPassword
        - oldPassword
    UserEditInfoRequest:
      type: object
      properties:
        nickname:
          type: string
          minLength: 1
      required:
        - nickname
    PasswordForgetEditRequest:
      type: object
      properties:
        mail:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 1
        confirmPassword:
          type: string
          minLength: 1
      required:
        - code
        - confirmPassword
        - mail
        - newPassword
    UserRegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        confirmPassword:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        nickname:
          type: string
          minLength: 1
        avatar:
          type: string
      required:
        - confirmPassword
        - email
        - nickname
        - password
        - username
    LoginCmd:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        isRemember:
          type: boolean
      required:
        - password
        - username
    ResultSysUserResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SysUserResponse'
        code:
          type: integer
          format: int32
        msg:
          type: string
    SysUserResponse:
      type: object
      description: 用户
      properties:
        id:
          type: string
          description: id
        username:
          type: string
          description: 用户名
        nickname:
          type: string
          description: 昵称
        avatar:
          type: string
          description: 头像
        email:
          type: string
          description: 邮箱地址
        status:
          type: integer
          format: int32
          description: 状态
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        lastLoginAt:
          type: string
          format: date-time
          description: 最后登录时间
  responses: {}
  securitySchemes: {}
servers: []
security: []